requirements.txt
aniso8601==7.0.0
appnope==0.1.0
asgiref==3.2.5
backcall==0.1.0
decorator==4.4.2
Django==3.0.4
django-cors-headers==3.2.1
django-extensions==2.2.8
django-filter==2.2.0
django-graphql-auth==0.3.6
django-graphql-jwt==0.3.0
django-sendgrid-v5==0.8.1
future==0.18.2
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.3.1
graphql-relay==2.0.1
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
nodeenv==1.3.5
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
promise==2.3
prompt-toolkit==3.0.4
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
python-dotenv==0.12.0
python-http-client==3.2.6
pytz==2019.3
Rx==1.6.1
sendgrid==6.2.0
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.1.9
